requirements.txt
flask
azure-ai-projects
azure-identity
gunicorn

